refactor: split helpers into multiple util files
